Refactored HTTP handler for site data
ref https://linear.app/ghost/issue/AP-592

We don't want to define our HTTP handlers inline in app.ts, instead pulling
them into separate files, this will keep app.ts smaller and let use share code
between the handlers easier.
