+++ src/app.ts
@@ -78,6 +78,7 @@ import {
 } from './dispatchers';
 import {
     followAction,
+    getSiteDataHandler,
     inboxHandler,
     likeAction,
     noteAction,
@@ -576,25 +577,7 @@ if (queue instanceof GCloudPubSubPushMessageQueue) {
 app.get(
     '/.ghost/activitypub/site',
     requireRole(GhostRole.Owner),
-    async (ctx) => {
-        const request = ctx.req;
-        const host = request.header('host');
-        if (!host) {
-            ctx.get('logger').info('No Host header');
-            return new Response('No Host header', {
-                status: 401,
-            });
-        }
-
-        const site = await getSite(host, true);
-
-        return new Response(JSON.stringify(site), {
-            status: 200,
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        });
-    },
+    getSiteDataHandler,
 );
 
 app.use(async (ctx, next) => {
